travel_assistant_back/requirements.txt
beautifulsoup4>=4.12.2
chromadb>=0.4.13
duckduckgo-search>=3.9.3
wikipedia>=1.4.0
python-dotenv>=1.0.0
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.4.2
requests>=2.31.0
mangum>=0.17.0
google-generativeai==0.3.2
langchain>=0.1.7
langchain-community>=0.0.9
langchain-google-genai>=0.0.5
langchain-text-splitters>=0.0.1